scan_manager successfully moved to sub-module
